added helpers (really important)
nsFunctionCaller and helper functions allow a script to have dynamic memory. it creates a temp script that writes to a temp output file to send the results back to the "main" script. currently it works really well but it doesnt check to see if there is enough ram to dynamically allocate to the specific method call. TODO add that